+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Features
 
 * adc: added initialization mappings for pins ADC1_INP12 (PC2) and ADC1_INP13 (PC3) (Not accessible on Daisy Seed)
+* rng: added new RandomNumberGenerator module that provides access to the hardware True Random Number Generator
 
 ### Bug fixes